refactor: update Makefiles and refactor attribute enums in models package
